p2p/netutil: change AddrIP -> AddrAddr
